+++ tencentcloud/es/v20180416/errors.go
@@ -23,6 +23,9 @@ const (
 	// 操作未授权。
 	AUTHFAILURE_UNAUTHORIZEDOPERATION = "AuthFailure.UnAuthorizedOperation"
 
+	// 操作失败。
+	FAILEDOPERATION = "FailedOperation"
+
 	// 集群资源配额限制错误。
 	FAILEDOPERATION_CLUSTERRESOURCELIMITERROR = "FailedOperation.ClusterResourceLimitError"
 
@@ -34,6 +37,9 @@ const (
 
 	// 集群索引没有副本存在。
 	FAILEDOPERATION_ERRORCLUSTERSTATENOREPLICATION = "FailedOperation.ErrorClusterStateNoReplication"
+
+	// 集群状态不健康。
+	FAILEDOPERATION_ERRORCLUSTERSTATEUNHEALTH = "FailedOperation.ErrorClusterStateUnhealth"
 
 	// 账户未绑定信用卡或paypal，无法支付。
 	FAILEDOPERATION_NOPAYMENT = "FailedOperation.NoPayment"
@@ -64,6 +70,18 @@ const (
 
 	// 资源不存在。
 	RESOURCENOTFOUND = "ResourceNotFound"
+
+	// 资源不可用。
+	RESOURCEUNAVAILABLE = "ResourceUnavailable"
+
+	// 资源售罄。
+	RESOURCESSOLDOUT = "ResourcesSoldOut"
+
+	// 未授权操作。
+	UNAUTHORIZEDOPERATION = "UnauthorizedOperation"
+
+	// 未知参数错误。
+	UNKNOWNPARAMETER = "UnknownParameter"
 
 	// 操作不支持。
 	UNSUPPORTEDOPERATION = "UnsupportedOperation"